# caseframesvc/swagger.yml
swagger: '2.0'
info:
  title: Case Frame Service API
  description: Provide access to Address Frame and Response Management information
  version: 8.11.0
schemes:
  - http
basePath: /frameservice
produces:
  - application/json
...
paths:
...
  /addresses/{uprn}:
    get:
      summary:  Get the address information for the specified UPRN.
      description: Return the full address details for the supplied Unique Property Reference Number (UPRN).
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
        - name: uprn
          description: Unique Property Reference Number for which to return address details.
          in: path
          required: true
          type: integer
      tags:
        - Frame
      responses:
        '200':
          description: Address details for supplied UPRN.
          schema:
            $ref: '#/definitions/Address'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /addresses/postcode/{postcode}:
    get:
      summary:  Get address information for a postcode.
      description: Return all address details for the supplied postcode.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
        - name: postcode
          description: Postal code.
          in: path
          required: true
          type: string
      tags:
        - Frame
      responses:
        '200':
          description: Address details for supplied postcode.
          schema:
            type: array
            items:
              $ref: '#/definitions/Address'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /cases/uprn/{uprn}:
    get:
      summary:  Get the case information for the specified UPRN.
      description: Return details of cases for the supplied Unique Property Reference Number (UPRN).
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
        - name: uprn
          description: Unique Property Reference Number for which to return details of cases.
          in: path
          required: true
          type: integer
      tags:
        - Response
      responses:
        '200':
          description: Case details for supplied UPRN.
          schema:
            type: array
            items:
              $ref: '#/definitions/Case'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /cases/questionnaire/{questionnaireid}:
    get:
      summary:  Get the case information for the specified questionnaire.
      description: Return details of a case for the supplied questionnaireid.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
        - name: questionnaireid
          description: Unique questionnaire Id.
          in: path
          required: true
          type: integer
      tags:
        - Response
      responses:
        '200':
          description: Case details for supplied questionnaireid.
          schema:
              $ref: '#/definitions/Case'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /cases/{caseid}:
    get:
      summary:  Get the case information for the specified case.
      description: Return details of a case for the supplied caseid.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
        - name: caseid
          description: Unique case Id.
          in: path
          required: true
          type: integer
      tags:
        - Response
      responses:
        '200':
          description: Case details for supplied caseid.
          schema:
              $ref: '#/definitions/Case'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /cases/{caseid}/events:
    get:
      summary:  Get the events that occurred on a specified case.
      description: Return all events for the supplied caseid.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
        - name: caseid
          description: Unique case Id.
          in: path
          required: true
          type: integer
      tags:
        - Response
      responses:
        '200':
          description: All event details for supplied caseid.
          schema:
            type: array
            items:
              $ref: '#/definitions/Event'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  /questionnaires/case/{caseid}:
    get:
      summary:  Get all questionnaire details for a specified case.
      description: |
        Return all questionnaire details for the supplied caseid.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
        - name: caseid
          description: Unique case Id.
          in: path
          required: true
          type: string
      tags:
        - Response
      responses:
        '200':
          description: All questionnaire details for a supplied caseid.
          schema:
            type: array
            items:
              $ref: '#/definitions/Questionnaire'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  /actions/{actionid}:
    get:
      summary:  Get the action details for a specified action.
      description: |
        Return the action details for the supplied actionid.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
        - name: actionid
          description: Unique action Id.
          in: path
          required: true
          type: integer
      tags:
        - Response
      responses:
        '200':
          description: Details for a single action.
          schema:
              $ref: '#/definitions/Action'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /actions/case/{caseid}:
    get:
      summary:  Get all action details for a specified case.
      description: |
        Return all action details for the supplied caseid.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
        - name: caseid
          description: Unique case Id.
          in: path
          required: true
          type: integer
      tags:
        - Response
      responses:
        '200':
          description: All action details for a supplied caseid.
          schema:
            type: array
            items:
              $ref: '#/definitions/Action'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  /casetypes:
    get:
      summary: Get all case types
      description: Return all case types.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
      tags:
        - Response
      responses:
        '200':
          description: An array of case types
          schema:
            type: array
            items:
              $ref: '#/definitions/CaseType'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /casetypes/{casetypeid}:
    get:
      summary: Get information for the specified case type.
      description: Return information for the supplied case type code.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
        - name: casetypeid
          in: path
          description: Unique case type Id.
          required: true
          type: integer
      tags:
        - Response
      responses:
        '200':
          description: A single case type.
          schema:
              $ref: '#/definitions/CaseType'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  /actionplans:
    get:
      summary: Get all action plans.
      description: Return details of all action plans.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
      tags:
        - Response
      responses:
        '200':
          description: An array of action plans.
          schema:
            type: array
            items:
              $ref: '#/definitions/ActionPlan'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /actionplans/{actionplanid}:
    get:
      summary: Get information for the specified action plan.
      description: Return information for the supplied action plan code.
      parameters:
        - name: Accept
          in: header
          description: Response type, application/json.
          required: false
          type: string
        - name: actionplanid
          in: path
          description: Unique action plan Id.
          required: true
          type: integer
      tags:
        - Response
      responses:
        '200':
          description: A single action plan.
          schema:
              $ref: '#/definitions/ActionPlan'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
definitions:
...
  Address:
    type: object
    properties:
      uprn:
        type: integer
        description: Unique Property Reference Number.
      addressType:
        type: string
        description: Address type CE, HH, I.
      estabType:
        type: string
        description: Communal Establishment type.
      addressLine1:
        type: string
        description: Address details.
      addressLine2:
        type: string
        description: Address details.
      townName:
        type: string
        description: Postal town name.
      postcode:
        type: string
        description: Postal Code.
      outputArea:
        type: string
        description: Census Geography Output Area code.
      lsoaArea:
        type: string
        description: Lower Layer Super Output Area code.
      msoaArea:
        type: string
        description: Middle layer Super Output Area code.
      ladCode:
        type: string
        description: GSS nine character LAD code.
      regionCode:
        type: string
        description: GSS nine character Region geography code.
      eastings:
        type: integer
        description: Easting or x-coordinate.
      northings:
        type: integer
        description: Northing or y-coordinate.
      htc:
        type: integer
        description: Hard to count code.
      latitude:
        type: number
        description: Position North or South of Equator.
      longitude:
        type: number
        description: Position East or West of Greenwich Meridian.
      ladName:
        type: string
        description: Local Authority Distict name.
  Case:
    type: object
    properties:
      caseId:
        type: integer
        description: Unique case identifier.
      uprn:
        type: integer
        description: Unique Property Reference Number.
      caseStatus:
        type: string
        description: Status of case.
      caseTypeId:
        type: integer
        description: Case Type Id
      createdDateTime:
        type: string
        format: date-time
        description: Creation date-time.
      createdBy:
        type: string
        description: Creator of case.
      sampleId:
        type: integer
        description: Sample Id.
      actionPlanId:
        type: integer
        description: Action plan for case.
      surveyId:
        type: integer
        description:  Survey instrument.
      parentCaseId:
        type: integer
        description: Linked parent case.
      questionSet:
        type: string
        description: Type of questionnaire set.
  Event:
    type: object
    properties:
      caseEventId:
        type: integer
        description: Unique case event identifier.
      caseId:
        type: integer
        description: Unique case identifier.
      description:
        type: string
        description: Case event note.
      createdDateTime:
        type: string
        format: date-time
        description: Creation date-time.
      createdBy:
        type: string
        description: Creator of event note.
      category:
        type: string
        description: Category.
  Questionnaire:
    type: object
    properties:
      questionnaireId:
        type: integer
        description: Unique questionnaire ID.
      caseId:
        type: integer
        description: Unique case identifier.
      questionnaireStatus:
        type: string
        description: Questionnaire status.
      dispatchDateTime:
        type: string
        format: date-time
        description: Date and time the questionnaire was dispatched.
      responseDateTime:
        type: string
        format: date-time
        description: Date and time the questionnaire was sent for return.
      receiptDateTime:
        type: string
        format: date-time
        description: Date and time the questionnaire was recieved.
      questionSet:
        type: string
        description: Type of questionnaire set.
      iac:
        type: string
        description: Internet Access Code.
  Action:
    type: object
    properties:
      actionId:
        type: integer
        description: Unique action ID.
      caseId:
        type: integer
        description: Unique case identifier.
      actionPlanId:
        type: integer
        description: Unique action plan Id.
      actionStatus:
        type: string
        description: Action status.
      actionType:
        type: string
        description: Type of action.
      priority:
        type: string
        description: The priority of the Action
      situation:
        type: string
        description: Field status, e.g. allocated, unallocated.
      createdDateTime:
        type: string
        format: date-time
        description: Date and time the action was created.
      createdBy:
        type: string
        description: Creator of action.
...
  CaseType:
    type: object
    properties:
      caseTypeId:
        type: integer
        description: Unique Id of case type.
      caseTypeName:
        type: string
        description: Name of case type.
      description:
        type: string
        description: Description of case type.
      actionPlanId:
        type: integer
        description: Unique action plan Id.
      questionSet:
        type: string
        description: Type of questionnaire set.
...
  ActionPlan:
    type: object
    properties:
      actionPlanId:
        type: integer
        description: Unique Id of action plan.
      actionPlanName:
        type: string
        description: Name of action plan.
      description:
        type: string
        description: Description of action plan.
      rules:
        type: string
        description: Action plan steps.
...
  Error:
    type: object
    properties:
      code:
        type: string
      timestamp:
        type: string
      message:
        type: string